fix(Avatar): set max width for avatar to avoid ellipse shape
When a min font size is set inside browser the proportion is not respected
and the avatar results in extending its width to include the text content.
By setting the max-width equal to the min-width the shape should be maintained.

Change also the calculation for the dimensions of the icon to be supported by Safari

refs: CDS-77