fix: change how the fonts are imported and bundled
Import fonts directly so that rollup is able to recognize them
and bundle with the url-plugin as inline resources.
By bundling them this way, fonts are resolved and used when setting
the loadDefaultFont prop in the ThemeProvider.

Update also webpack to recognize the fonts as asset/resource for the styleguide.

Add a file transformer for jest and update the config to use it.

Remove no more used plugin from dependencies.

refs: CDS-110