feat(ListV2): create ListV2 component
Create a new implementation of the list, which is a refactored version.
The items of the list are now passed as children (so as components) and must be
defined outside.
The management of the selected and active items are now totally left up to the external
usege, removing the hibrid solution implemented in the previous version of the List.
Refer to the documentation for more information and examples.

refs: CDS-107