fix(CustomModal): do not prevent default behavior of click inside modal
By blinldy preventing default behavior on click, basic html elements
like the anchor, in which the click is left to be the default one,
stop working.

Remove the prevention of the default behaviour from the custom modal,
and leave the external user in charge of preventing what needs to be prevented.

refs: CDS-102