fix(ChipInput): scroll to the end while creating chip
While creating chips, scroll the container both horizontally
and vertically to make the last chip and the input always visible.
The scroll is now made each time the items change, meaning that
the scroll is performed even when the items change because of
a paste or because of a programmatic/controlled usage.

Refs: CDS-146