fix: avoid styled components props propagation to html
Prefix styled component props to prevent propagation of invalid attributes to HTML.

Warning: with this change, the disabled prop is no more implicitly passed down to those elements which don't expect it (e.g. the div).
This means that tests that are checking the existence of the HTML attribute disabled won't work anymore.
Replace the assertions with the toBeEnabled/toBeDisabled where possible. Where not possible, the component probably requires a refactor to use the right HTML element.

Refs: CDS-144 (#316)