fix(Dropdown): prevent onClick of parent item to be called on child click
Prevent the click callback of the parent item to be called when
a child is clicked.
This align the behavior between the click fired from a mouse event
and the one fired by a keyboard event and stop unwanted chain of invocations.

Update all keyboard events to use keydown instead of keypress since
the keypress is deprecated and stopped working on some browser.

Refs: CDS-189